+++ src/test/scala/au/com/agl/arc/extract/ParquetExtractSuite.scala
@@ -68,7 +68,8 @@ class ParquetExtractSuite extends FunSuite with BeforeAndAfter {
         params=Map.empty,
         persist=false,
         numPartitions=None,
-        partitionBy=Nil
+        partitionBy=Nil,
+        contiguousIndex=None
       )
     )
 
@@ -108,7 +109,8 @@ class ParquetExtractSuite extends FunSuite with BeforeAndAfter {
         params=Map.empty,
         persist=false,
         numPartitions=None,
-        partitionBy=Nil
+        partitionBy=Nil,
+        contiguousIndex=None
       )
     )
     assert(spark.catalog.isCached(outputView) === false)
@@ -124,7 +126,8 @@ class ParquetExtractSuite extends FunSuite with BeforeAndAfter {
         params=Map.empty,
         persist=true,
         numPartitions=None,
-        partitionBy=Nil
+        partitionBy=Nil,
+        contiguousIndex=None
       )
     )
     assert(spark.catalog.isCached(outputView) === true)     
@@ -161,7 +164,8 @@ class ParquetExtractSuite extends FunSuite with BeforeAndAfter {
           params=Map.empty,
           persist=false,
           numPartitions=None,
-          partitionBy=Nil
+          partitionBy=Nil,
+          contiguousIndex=None
         )
       )
     }
@@ -179,7 +183,8 @@ class ParquetExtractSuite extends FunSuite with BeforeAndAfter {
           params=Map.empty,
           persist=false,
           numPartitions=None,
-          partitionBy=Nil
+          partitionBy=Nil,
+          contiguousIndex=None
         )
       )
     }
@@ -196,7 +201,8 @@ class ParquetExtractSuite extends FunSuite with BeforeAndAfter {
         params=Map.empty,
         persist=false,
         numPartitions=None,
-        partitionBy=Nil
+        partitionBy=Nil,
+        contiguousIndex=None
       )
     )